ダメージ予想関連の関数をリファクタリングし、非推奨の関数を削除。新しい関数を追加して戦闘結果を取得し、ダメージ予想を計算するロジックを整理。
