+++ src/js/td-scenes/battle/animation/game-state/input-command/index.ts
@@ -3,10 +3,12 @@ import { GameStateX, InputCommand } from "gbraver-burst-core";
 import { all } from "../../../../../animation/all";
 import { Animate } from "../../../../../animation/animate";
 import { empty } from "../../../../../animation/delay";
+import { getEnableMaxBattery } from "../../../get-enable-max-battery";
+import { getInitialBattery } from "../../../get-initial-battery";
+import { calcPredicatedDamage, getBattleResultForPredicatedDamage } from "../predicated-damage";
 import { StateAnimationProps } from "../state-animation-props";
 import { activeArmdozerSprite } from "./active-armdozer-sprite";
 import { showCommand } from "./show-command";
-import { showPredicatedDamage } from "./show-predicated-damage";
 import { updateGauge } from "./update-gauge";
 
 /**
@@ -38,6 +40,19 @@ export function inputCommandAnimation(
   }
 
   const isPlayerTurn = playerId === activePlayerId;
+  const enableMaxBattery = getEnableMaxBattery(playerCommand.command);
+  const playerBattery =
+      controllerType === "BigButton"
+        ? getInitialBattery(enableMaxBattery)
+        : player.armdozer.battery;
+  const battleResult = getBattleResultForPredicatedDamage({
+    attacker,
+    defender,
+    playerId,
+    playerBattery,
+  });
+  const predicatedDamage = calcPredicatedDamage(battleResult);
+  defenderHUD.predicatedDamage.set(predicatedDamage);
   return all(
     updateGauge(view.hud.players, players),
     showCommand({
@@ -50,13 +65,6 @@ export function inputCommandAnimation(
     view.hud.gameObjects.timeScaleButton.open(animationTimeScale),
     ...view.hud.players.map((p) => p.statusIcon.open()),
     activeArmdozerSprite(view.td.armdozers, activePlayerId),
-    showPredicatedDamage({
-      attacker,
-      defender,
-      defenderHUD,
-      playerId,
-      commands: playerCommand.command,
-      controllerType,
-    }),
+    defenderHUD.predicatedDamage.show(predicatedDamage),
   );
 }
