+++ src/js/td-scenes/battle/animation/game-state/predicated-damage.ts
