--- src/js/td-scenes/battle/procedure/do-battle-simulation.ts
@@ -1,4 +1,0 @@
-export const doBattleSimulation = (props: {
-}) => {
-
-}
